Bot: Reorganize DTOs
