+++ bot/src/main/java/edu/java/bot/controller/BotController.java
@@ -1,6 +1,6 @@
 package edu.java.bot.controller;
 
-import edu.java.bot.dto.LinkUpdate;
+import edu.java.bot.dto.request.LinkUpdateRequest;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Log4j2
 public class BotController {
     @GetMapping("/updates")
-    void processUpdate(LinkUpdate update) {
+    void processUpdate(LinkUpdateRequest update) {
        log.info("Received updates: {}", update);
     }
 }
