--- bot/src/main/java/edu/java/bot/dto/LinkUpdate.java
@@ -1,7 +1,0 @@
-package edu.java.bot.dto;
-
-import java.net.URI;
-import java.util.List;
-
-public record LinkUpdate(long id, URI url, String description, List<Long> tgChatIds) {
-}
