+++ bot/src/main/java/edu/java/bot/dto/request/LinkUpdateRequest.java
@@ -1,0 +1,7 @@
+package edu.java.bot.dto.request;
+
+import java.net.URI;
+import java.util.List;
+
+public record LinkUpdateRequest(long id, URI url, String description, List<Long> tgChatIds) {
+}
