+++ bot/src/main/java/edu/java/bot/dto/request/RemoveLinkRequest.java
@@ -1,4 +1,4 @@
-package edu.java.bot.dto;
+package edu.java.bot.dto.request;
 
 import java.net.URI;
