+++ bot/src/main/java/edu/java/bot/dto/response/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package edu.java.bot.dto;
+package edu.java.bot.dto.response;
 
 import java.util.Arrays;
 import java.util.List;
