+++ bot/src/main/java/edu/java/bot/dto/response/LinkResponse.java
@@ -1,4 +1,4 @@
-package edu.java.bot.dto;
+package edu.java.bot.dto.response;
 
 import java.net.URI;
